consle tree view for SHS
